solutions/mod-03/requirements/flask-sqlalchemy-mysql.txt
flask
flask-sqlalchemy
sqlalchemy-utils
pymysql
cryptography
